Remove console logs and unused imports for production readiness
